# Conflicts:
#	applications/minotari_console_wallet/src/ui/components/register_template_tab.rs
#	applications/minotari_console_wallet/src/ui/state/tasks.rs

+++ base_layer/core/src/chain_storage/lmdb_db/lmdb_db.rs
@@ -472,8 +472,8 @@ impl LMDBDatabase {
                         &MetadataValue::HorizonData(horizon_data.clone()),
                     )?;
                 },
-                InsertBadBlock { hash, height } => {
-                    self.insert_bad_block_and_cleanup(&write_txn, hash, *height)?;
+                InsertBadBlock { hash, height, reason } => {
+                    self.insert_bad_block_and_cleanup(&write_txn, hash, *height, reason.to_string())?;
                 },
                 InsertReorg { reorg } => {
                     lmdb_replace(&write_txn, &self.reorgs, &reorg.local_time.timestamp(), &reorg)?;
@@ -889,6 +889,7 @@ impl LMDBDatabase {
             .fetch_height_from_hash(write_txn, block_hash)
             .or_not_found("Block", "hash", hash_hex)?;
         let next_height = height.saturating_add(1);
+        let prev_height = height.saturating_sub(1);
         if self.fetch_block_accumulated_data(write_txn, next_height)?.is_some() {
             return Err(ChainStorageError::InvalidOperation(format!(
                 "Attempted to delete block at height {} while next block still exists",
@@ -905,6 +906,20 @@ impl LMDBDatabase {
         let mut smt = self.fetch_tip_smt()?;
 
         self.delete_block_inputs_outputs(write_txn, block_hash.as_slice(), height, &mut smt)?;
+
+        let new_tip_header = self.fetch_chain_header_by_height(prev_height)?;
+        let root = FixedHash::try_from(smt.hash().as_slice())?;
+        if root != new_tip_header.header().output_mr {
+            error!(
+                target: LOG_TARGET,
+                "Deleting block, new smt root(#{}) did not match expected (#{}) smt root",
+                    root.to_hex(),
+                    new_tip_header.header().output_mr.to_hex(),
+            );
+            return Err(ChainStorageError::InvalidOperation(
+                "Deleting block, new smt root did not match expected smt root".to_string(),
+            ));
+        }
         self.insert_tip_smt(write_txn, &smt)?;
         self.delete_block_kernels(write_txn, block_hash.as_slice())?;
 
@@ -955,7 +970,17 @@ impl LMDBDatabase {
                 continue;
             }
             let smt_key = NodeKey::try_from(utxo.output.commitment.as_bytes())?;
-            output_smt.delete(&smt_key)?;
+            match output_smt.delete(&smt_key)? {
+                DeleteResult::Deleted(_value_hash) => {},
+                DeleteResult::KeyNotFound => {
+                    error!(
+                        target: LOG_TARGET,
+                        "Could not find input({}) in SMT",
+                        utxo.output.commitment.to_hex(),
+                    );
+                    return Err(ChainStorageError::UnspendableInput);
+                },
+            };
             lmdb_delete(
                 txn,
                 &self.utxo_commitment_index,
@@ -1006,8 +1031,15 @@ impl LMDBDatabase {
                 utxo_mined_info.output.minimum_value_promise,
             );
             let smt_key = NodeKey::try_from(input.commitment()?.as_bytes())?;
-            let smt_node = ValueHash::try_from(input.smt_hash(row.spent_height).as_slice())?;
-            output_smt.insert(smt_key, smt_node)?;
+            let smt_node = ValueHash::try_from(input.smt_hash(utxo_mined_info.mined_height).as_slice())?;
+            if let Err(e) = output_smt.insert(smt_key, smt_node) {
+                error!(
+                    target: LOG_TARGET,
+                    "Output commitment({}) already in SMT",
+                    input.commitment()?.to_hex(),
+                );
+                return Err(e.into());
+            }
 
             trace!(target: LOG_TARGET, "Input moved to UTXO set: {}", input);
             lmdb_insert(
@@ -1224,7 +1256,14 @@ impl LMDBDatabase {
             if !output.is_burned() {
                 let smt_key = NodeKey::try_from(output.commitment.as_bytes())?;
                 let smt_node = ValueHash::try_from(output.smt_hash(header.height).as_slice())?;
-                output_smt.insert(smt_key, smt_node)?;
+                if let Err(e) = output_smt.insert(smt_key, smt_node) {
+                    error!(
+                        target: LOG_TARGET,
+                        "Output commitment({}) already in SMT",
+                        output.commitment.to_hex(),
+                    );
+                    return Err(e.into());
+                }
             }
 
             let output_hash = output.hash();
@@ -1260,7 +1299,14 @@ impl LMDBDatabase {
             let smt_key = NodeKey::try_from(input_with_output_data.commitment()?.as_bytes())?;
             match output_smt.delete(&smt_key)? {
                 DeleteResult::Deleted(_value_hash) => {},
-                DeleteResult::KeyNotFound => return Err(ChainStorageError::UnspendableInput),
+                DeleteResult::KeyNotFound => {
+                    error!(
+                        target: LOG_TARGET,
+                        "Could not find input({}) in SMT",
+                        input_with_output_data.commitment()?.to_hex(),
+                    );
+                    return Err(ChainStorageError::UnspendableInput);
+                },
             };
 
             let features = input_with_output_data.features()?;
@@ -1554,13 +1600,14 @@ impl LMDBDatabase {
         txn: &WriteTransaction<'_>,
         hash: &HashOutput,
         height: u64,
+        reason: String,
     ) -> Result<(), ChainStorageError> {
         #[cfg(test)]
         const CLEAN_BAD_BLOCKS_BEFORE_REL_HEIGHT: u64 = 10000;
         #[cfg(not(test))]
         const CLEAN_BAD_BLOCKS_BEFORE_REL_HEIGHT: u64 = 0;
 
-        lmdb_replace(txn, &self.bad_blocks, hash.deref(), &height)?;
+        lmdb_replace(txn, &self.bad_blocks, hash.deref(), &(height, reason))?;
         // Clean up bad blocks that are far from the tip
         let metadata = fetch_metadata(txn, &self.metadata_db)?;
         let deleted_before_height = metadata
@@ -1570,8 +1617,9 @@ impl LMDBDatabase {
             return Ok(());
         }
 
-        let num_deleted =
-            lmdb_delete_each_where::<[u8], u64, _>(txn, &self.bad_blocks, |_, v| Some(v < deleted_before_height))?;
+        let num_deleted = lmdb_delete_each_where::<[u8], (u64, String), _>(txn, &self.bad_blocks, |_, (v, _)| {
+            Some(v < deleted_before_height)
+        })?;
         debug!(target: LOG_TARGET, "Cleaned out {} stale bad blocks", num_deleted);
 
         Ok(())
@@ -2356,9 +2404,22 @@ impl BlockchainBackend for LMDBDatabase {
             .collect()
     }
 
-    fn bad_block_exists(&self, block_hash: HashOutput) -> Result<bool, ChainStorageError> {
+    fn bad_block_exists(&self, block_hash: HashOutput) -> Result<(bool, String), ChainStorageError> {
         let txn = self.read_transaction()?;
-        lmdb_exists(&txn, &self.bad_blocks, block_hash.deref())
+        // We do this to ensure backwards compatibility on older exising dbs that did not store a reason
+        let exist = lmdb_exists(&txn, &self.bad_blocks, block_hash.deref())?;
+        match lmdb_get::<_, (u64, String)>(&txn, &self.bad_blocks, block_hash.deref()) {
+            Ok(Some((_height, reason))) => Ok((true, reason)),
+            Ok(None) => Ok((false, "".to_string())),
+            Err(ChainStorageError::AccessError(e)) => {
+                if exist {
+                    Ok((true, "No reason recorded".to_string()))
+                } else {
+                    Err(ChainStorageError::AccessError(e))
+                }
+            },
+            Err(e) => Err(e),
+        }
     }
 
     fn clear_all_pending_headers(&self) -> Result<usize, ChainStorageError> {
@@ -2480,7 +2541,7 @@ fn fetch_metadata(txn: &ConstTransaction<'_>, db: &Database) -> Result<ChainMeta
         fetch_pruned_height(txn, db)?,
         fetch_accumulated_work(txn, db)?,
         fetch_best_block_timestamp(txn, db)?,
-    ))
+    )?)
 }
 
 // Fetches the chain height from the provided metadata db.
